// src/app/edit-article/types/editArticleState.interface.ts
import { BackendErrorsInterface } from '../../shared/types/backendErrors.interface';
import { ArticleInterface } from '../../shared/types/article.interface';

export interface EditArticleStateInterface {
  article: ArticleInterface | null;
  isSubmitting: boolean;
  validationErrors: BackendErrorsInterface | null;
  isLoading: boolean;
}